+++ src/proxy/routerGen.go
@@ -5,6 +5,7 @@ import (
 	"regexp"
 	"strconv"
 	"time"
+	"net/url"
 
 	"github.com/azukaar/cosmos-server/src/user"
 	"github.com/azukaar/cosmos-server/src/constellation"
@@ -93,6 +94,17 @@ func RouterGen(route utils.ProxyRouteConfig, router *mux.Router, destination htt
 
 	if route.UseHost {
 		origin = origin.Host(route.Host)
+
+		// if Scheme is not http/https, discard
+		urlRoute, err := url.Parse(route.Target)
+		if err != nil {
+			utils.Error("Invalid target URL: "+route.Target, err)
+			return nil
+		}
+
+		if urlRoute.Scheme != "http" && urlRoute.Scheme != "https" {
+			return nil
+		}
 	}
 	
 	if route.UsePathPrefix {